requirements.txt
# Excel RAG System Dependencies

# Core
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0

# Vector Store
chromadb>=0.4.0

# LLM Client (Nebius - OpenAI compatible)
openai>=1.0.0

# Google Gemini (backup, currently paused)
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
